finstat_system_vscode/requirements.txt
pandas>=2.2.0
dbfread>=2.0.7
openpyxl>=3.1.2
XlsxWriter>=3.1.9
pyyaml>=6.0.1
tqdm>=4.66.4
SQLAlchemy>=2.0.29
rarfile>=4.0
openai>=1.35.0
python-dotenv>=1.0.1
